v18: time_proj'u onceden hesaplanmis tabloya cevir, grafta Gather birak
local-dream kaynagi (QnnModel.hpp executeUnetGraphs) kesinlestirdi: motor
timestamp'i ham int32 yaziyor (positionData[0] = timestep), sample ve
text_embedding'i uint16_t* olarak yaziyor. Yani timestamp graf sinirinda
INT_32 olmak ZORUNDA — v17'nin float32 sinir denemesi bu yuzden bos.

HTP'nin kabul listesinde int32 girdiyi kuantize ciktiya baglayan hicbir sekil
islemi yok, ama Gather var: veri kuantize / indeks int32 / cikti kuantize.
time_proj yalnizca t'nin fonksiyonu ve t tamsayi oldugundan tum timestep'ler
onceden hesaplanip [num_train_timesteps, 320] sabit tabloya konuyor; grafta
Gather kaliyor. Sonuc birebir ayni.

- 01_export_onnx.py: TimeProjTable (cozunurluk basina tekrar sarmalamaz),
  timestep ornegi tekrar int32 (export v18)
- gen_io_config.py: timestamp Src=int32 / Desired=int32
- 02_gen_quant_data.py: timestep raw'lari tekrar int32 (calib v4)
- convert_all.sh: EXPORT_VERSION=18, CALIB_VERSION=4
- docs/DURUM.md: kaynak kod alintilari, referans paket icerigi, gerekce

+++ scripts/gen_io_config.py
@@ -108,7 +108,8 @@ def main() -> None:
         "",
         "Input Tensor Configuration:",
         _tensor_block("sample", args.dtype, 1, "Input", q_sample),
-        _tensor_block("timestamp", "int32", 2, "Input"),
+        # ONNX'te de int32: grafta Gather indeksi, kuantize edilmiyor.
+        _tensor_block("timestamp", "int32", 2, "Input", src_dtype="int32"),
         _tensor_block("text_embedding", args.dtype, 3, "Input", q_text),
         "",
         "Output Tensor Configuration:",
